% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_UN_country_tmap.R
\name{create_UN_country_tmap}
\alias{create_UN_country_tmap}
\title{create_UN_country_tmap}
\usage{
create_UN_country_tmap(iso3, preview = FALSE)
}
\arguments{
\item{iso3}{country code}

\item{preview}{preview as dynamic map}
}
\value{
an object of class \link[tmap]{tmap}
}
\description{
Creates a UN country tmap
}
\author{
Arturo Muñoz Albero
}
