% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdi4R.R
\docType{package}
\name{fdi4R}
\alias{fdi4R}
\title{Tools for data processing in support of fisheries data interoperability operations}
\description{
fdi4R provides a set of utility tools to do data processing in support of fisheries data interoperability operations
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
