% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_WoRMS_AphiaIDs.R
\name{get_WoRMS_AphiaID}
\alias{get_WoRMS_AphiaID}
\title{get_WoRMS_AphiaID}
\usage{
get_WoRMS_AphiaID(name)
}
\arguments{
\item{name}{a species scientific name}
}
\value{
the corresponding AphiaID if any, \code{NA} otherwise
}
\description{
Tries to get a WoRMS Aphia ID from a WoRMS using taxa matching. 
The function is wrapper on top of \link[worrms]{wm_records_taxamatch} function.
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
