% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_WoRMS_AphiaIDs.R
\name{get_WoRMS_AphiaIDs}
\alias{get_WoRMS_AphiaIDs}
\title{get_WoRMS_AphiaIDs}
\usage{
get_WoRMS_AphiaIDs(
  x,
  parallel = FALSE,
  parallel_handler = NULL,
  cl = NULL,
  ...
)
}
\arguments{
\item{x}{a vector of species scientific names}
}
\value{
a vector of the corresponding AphiaIDs if any, \code{NA} otherwise
}
\description{
Tries to get WoRMS Aphia IDs for a series of scientific names. The function
supports \code{parallel} mode for fast taxa matching.
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
